requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
Booktype==1.5
boto3==1.40.3
botocore==1.40.3
Django==5.2.4
ecdsa==0.19.1
fastapi==0.116.1
idna==3.10
jmespath==1.0.1
passlib==1.7.4
pdf2image==1.17.0
pillow==11.3.0
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-jose==3.5.0
rsa==4.9.1
s3transfer==0.13.1
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlparse==0.5.3
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
